fix(gpu): apply VkResult cast + GLFW_INCLUDE_VULKAN + gtest_main everywhere
Picks up three corrections that were applied to main.cpp and tests/CMakeLists
locally and propagates the VkResult-cast fix to every other call site:

- VkResult is an enum, not an integral type, so std::to_string(VkResult) is
  ill-formed under strict C++23. Cast to int at every error-formatting site
  in engine/src/gpu/{instance,device,swapchain,physical_device}.cpp.
- main.cpp: GLFW_INCLUDE_VULKAN so glfwCreateWindowSurface sees the Vulkan
  types and the function is declared with the right signature.
- tests/CMakeLists.txt: link GTest::gtest_main so test executables don't
  need to define their own main().

+++ app/src/main.cpp
@@ -56,7 +56,7 @@ void install_default_observers() {
                                           nullptr, &raw); vr != VK_SUCCESS) {
         return std::unexpected(noted::make_error(
             noted::ErrorCode::gpu_surface_lost,
-            std::string{"glfwCreateWindowSurface failed: VkResult="} + std::to_string(vr)));
+            std::string{"glfwCreateWindowSurface failed: VkResult="} + std::to_string(static_cast<int>(vr))));
     }
     return raw;
 }
